+++ web/documentserver-example/java-spring/src/main/resources/static/scripts/jscript.js
@@ -114,7 +114,7 @@ if (typeof jQuery !== "undefined") {
 
         if (!formatManager.isAutoConvertible(posExt)) {
             jq("#step2").addClass("done").removeClass("current");
-            loadScripts();
+            onuploaded();
             return;
         }
 
@@ -170,35 +170,14 @@ if (typeof jQuery !== "undefined") {
                         jq("#hiddenFileName").val(response.filename);
                         jq("#uploadFileName").text(response.filename);
                         jq("#step2").addClass("done").removeClass("current");
-                        loadScripts();
+                        onuploaded();
                     }
                 }
             });
         }, 1000);
     };
 
-    var loadScripts = function () {
-        if (!jq("#mainProgress").is(":visible")) {
-            return;
-        }
-        jq("#step3").addClass("current");
-
-        if (jq("#loadScripts").is(":empty")) {
-            var urlScripts = jq("#loadScripts").attr("data-docs");
-            var frame = "<iframe id=\"iframeScripts\" width=1 height=1 style=\"position: absolute; visibility: hidden;\" ></iframe>";
-            jq("#loadScripts").html(frame);
-            document.getElementById("iframeScripts").onload = onloadScripts;
-            jq("#loadScripts iframe").attr("src", urlScripts);
-        } else {
-            onloadScripts();
-        }
-    };
-
-    var onloadScripts = function () {
-        if (!jq("#mainProgress").is(":visible")) {
-            return;
-        }
-        jq("#step3").addClass("done").removeClass("current");
+    var onuploaded = function () {
         jq("#beginView, #beginEmbedded").removeClass("disable");
 
         var fileName = jq("#hiddenFileName").val();
@@ -266,7 +245,7 @@ if (typeof jQuery !== "undefined") {
 
     jq(document).on("click", "#skipPass", function () {
         jq("#blockPassword").hide();
-        loadScripts();
+        onuploaded();
     });
 
     jq(document).on("click", "#beginEdit:not(.disable)", function () {